arreglado #11

+++ src/includes/payload.h
@@ -12,7 +12,7 @@
 
 typedef struct t_payload{
     uint8_t * content; //puntero a el mensaje a inyectar como array de bytes
-    uint8_t counter; //índice del byte actual en el array
+    uint32_t counter; //índice del byte actual en el array
     long size; //tamaño del array
 } t_payload;
